refactor(l2): remove eth_sender module
