+++ crates/l2/tests/tests.rs
@@ -18,7 +18,7 @@ use ethrex_l2_sdk::{
     get_erc1967_slot, git_clone, wait_for_transaction_receipt,
 };
 use ethrex_rpc::{
-    clients::eth::{EthClient, L1MessageProof, eth_sender::Overrides, from_hex_string_to_u256},
+    clients::eth::{EthClient, L1MessageProof, Overrides, from_hex_string_to_u256},
     types::{
         block_identifier::{BlockIdentifier, BlockTag},
         receipt::RpcReceipt,
