+++ crates/networking/rpc/clients/mod.rs
@@ -3,4 +3,4 @@ pub mod beacon;
 pub mod eth;
 
 pub use auth::{EngineClient, errors::EngineClientError};
-pub use eth::{EthClient, errors::EthClientError, eth_sender::Overrides};
+pub use eth::{EthClient, Overrides, errors::EthClientError};
